+++ rest_framework_simple_api_key/analytics/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AnalyticsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "rest_framework_simple_api_key.analytics"
+    label = "rest_framework_simple_api_key_analytics"