+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 [Unreleased]
 ------------
 
+-  Analytics and Monitoring (#52)
+
 [v1.0.2] - 2023-09-16
 ------------------